requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain-groq
pypdf
faiss-gpu
langchain_text_splitters
